remove duplicate code
point {package.json}.browser to an existing file
